% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinepatterns.R
\name{combinepatterns}
\alias{combinepatterns}
\title{Plot combined eigenvectors}
\usage{
combinepatterns(ssa, pcs)
}
\arguments{
\item{ssa}{an `ssa` class object}

\item{pcs}{a vector specifying which eigenvectors to group}
}
\value{
None
}
\description{
A useful technique in Singular Spectrum Analysis is to look at the summed groupings
of eigenvectors. These groupings can usually reveal the interframe signal.
}
\details{
Oscillating patterns that are close in eigenvalue magnitude usually represent a signal component that cannot be represented by a single sinusoid.
}
\examples{
obj=ssa(co2,18)
combinepatterns(obj,2:3)
}
