% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.ssa.R
\name{embed.ssa}
\alias{embed.ssa}
\title{Embed a series into a trajectory matrix}
\usage{
embed.ssa(series, window.length = 12)
}
\arguments{
\item{window.length}{}
}
\value{
a Hankel-type, trajectory matrix X
}
\description{
This is a helper function for the `math5973NivProj2` package.
}
