% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hankelize.R
\name{hankelize}
\alias{hankelize}
\title{Skew-diagonal averaging to create a Hankel matrix}
\usage{
hankelize(X)
}
\arguments{
\item{X}{trajectory matrix}
}
\description{
This is a helper function for the `math5973NivProj2` package.
}
