% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ssa.R
\name{print.ssa}
\alias{print.ssa}
\title{Singular Spectrum Analysis}
\usage{
\method{print}{ssa}(ssa)
}
\arguments{
\item{ssa}{an object of class \code{ssa}}
}
\value{
None
}
\description{
Given an object of class \code{ssa}, the \code{print} function prints to the command line in descending order the standard deviations accounted for by each principal component.
}
\examples{
obj=ssa(co2,18)
print(obj)
}
