% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.R
\name{reconstruct}
\alias{reconstruct}
\title{Reconstruct a series from an \code{ssa} object}
\usage{
reconstruct(ssa, pcs = c(1, 2), pct.var = NULL, inplace = T, plot = T)
}
\arguments{
\item{ssa}{object created from \code{ssa()} function.}

\item{pcs}{a vector of integers specifying the projection bases of principal components.}

\item{pct.var}{(optional) a number in (0.0,1.0] specifying the desired reconstruction accuracy.}

\item{inplace}{\code{logical} indicating whether the reconstructed series should modify the \code{ssa} object. \code{FALSE} returns the reconstructed series.}

\item{plot}{\code{logical} \code{TRUE} generates a plot of the reconstructed series with residuals for comparison.}
}
\value{
if \code{inplace=FALSE}, returns a list of the reconstructed trajectory matrix and series.
}
\description{
Projects the series onto chosen principal components, and performs skew-diagonal averaging on projected data.
}
\examples{
obj=ssa(co2,18)
reconstruct(obj,1:5,plot=T)
}
