% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa.R
\name{ssa}
\alias{ssa}
\title{Singular Spectrum Analysis}
\usage{
ssa(series, window.length = 12)
}
\arguments{
\item{series}{a time-series or spatial-series}

\item{window.length}{the interval length to consider within series}
}
\value{
an object of class \code{ssa}
}
\description{
This functon constructs an \code{ssa} class object to perform a singular spectrum analysis.
}
\examples{
ssa(co2,18)
}
