% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ssa.R
\name{plot.ssa}
\alias{plot.ssa}
\title{Singular Spectrum Analysis}
\usage{
\method{plot}{ssa}(ssa, pcs = 1:4)
}
\arguments{
\item{ssa}{an object of class \code{ssa}}

\item{pcs}{a vector of integers specifying which principal components to plot}
}
\value{
None
}
\description{
A \code{ssa} class object to perform a singular spectrum analysis.
}
