% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ssa.R
\name{print.ssa}
\alias{print.ssa}
\title{Singular Spectrum Analysis}
\usage{
\method{print}{ssa}(ssa)
}
\arguments{
\item{ssa}{an object of class \code{ssa}}
}
\value{
an object of class \code{ssa}
}
\description{
A \code{ssa} class object to perform a singular spectrum analysis.
}
