% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ssa.R
\name{summary.ssa}
\alias{summary.ssa}
\title{Singular Spectrum Analysis}
\usage{
\method{summary}{ssa}(ssa)
}
\arguments{
\item{ssa}{an object of class \code{ssa}}
}
\value{
an object of class \code{ssa}
}
\description{
Given an \code{ssa} class object, the cumulative and total proportion of variance accounted for by each principal component is printed to the command line in descending order.
}
\details{
The sum of the eigenvalues of a covariance matrix is equal to the total variance in the data. The proportion of variance accounted for by a principal component is its eigenvalue divided by the sum of all eigenvalues.
}
\examples{
obj=ssa(co2,18)
summary(obj)
}
