% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowtest.R
\name{windowtest}
\alias{windowtest}
\title{Window length tests}
\usage{
windowtest(ssa)
}
\arguments{
\item{ssa}{}
}
\value{
None
}
\description{
Plots Lag-correlation, Autocorrelation, and Partial Autocorrelation as
a heuristic for determining optimal window length.
}
\examples{
obj=ssa(co2,18)
windowtest(obj)

}
